feat(deps): llm-kernel 0.1.0 → 0.1.1 — cosine f64 및 OpenAI 모델 디스패치 위임

+++ src/adapters/infra/search_engines.rs
@@ -299,11 +299,12 @@ pub fn sparse_entity_boost(counts: &HashMap<String, usize>, entity_id: &str) -> 
 // Cosine similarity
 // ---------------------------------------------------------------------------
 
-/// Compute cosine similarity between two f32 vectors.
+/// Compute cosine similarity between two f32 vectors, returning an f64 result.
 ///
-/// Delegates to [`llm_kernel::embedding::cosine_similarity`] (single-pass f32).
+/// Delegates to [`llm_kernel::embedding::cosine_similarity`] which performs
+/// f64-precision accumulation internally to avoid catastrophic cancellation.
 pub fn cosine_similarity(a: &[f32], b: &[f32]) -> f64 {
-    llm_kernel::embedding::cosine_similarity(a, b) as f64
+    llm_kernel::embedding::cosine_similarity(a, b)
 }
 
 // ---------------------------------------------------------------------------
